backend/requirements.txt
fastapi>=0.68.1
uvicorn>=0.15.0
sqlalchemy>=1.4.23
pydantic>=1.8.2
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.5
aiofiles>=0.7.0
python-dotenv>=0.19.0
PyJWT>=2.3.0
databases[sqlite]>=0.5.3
aiosqlite>=0.17.0
python-dateutil>=2.8.2
numpy>=1.24.0
pandas>=1.5.0
scikit-learn>=0.24.2  # For advanced analytics
requests>=2.26.0
httpx>=0.19.0
websockets>=10.0
